Moved editing of ingredients to seperate component
