+++ app/screens/Food/Detail.js
@@ -7,6 +7,7 @@ import { getCurrentTheme } from '../../redux/themeSlice';
 
 
 import Tag from '../../components/Tag';
+import EditIngredients from './EditIngredients';
 
 // UI library 
 import * as eva from '@eva-design/eva';
@@ -30,14 +31,6 @@ const servingsChoices = [
     {label: '10', value: 10},
 ];
 
-const units = [
-    'g',
-    'kg',
-    'l',
-    'dl',
-    'ml',
-    'st'
-];
 
 function FoodDetail({route, navigation}) {
     let recipe = route.params.route.params.recipe;
@@ -49,7 +42,8 @@ function FoodDetail({route, navigation}) {
     const [isEditing, setIsEditing] = useState(false);
 
     const [newTitle, setNewTitle] = useState(recipe.title);
-    const [newIngredients, setNewIngredients] = useState(recipe.ingredients);
+    const newIngredientsState = useState(recipe.ingredients);
+    const [newIngredients, setNewIngredients] = newIngredientsState;
 
     function ToggleIsEditing() {
         setIsEditing(!isEditing);
@@ -87,78 +81,9 @@ function FoodDetail({route, navigation}) {
     }
 
     function Ingredients() {
-        let defaultTextEdit = {editingIndex:-1, text:''}
-        // Used to keep track of which ingredient is being edited and only update newIngredients when the user is done editing
-        let [currEditAmount, setCurrentEditAmount] = useState(defaultTextEdit);
-        let [currEditUnit, setCurrentEditUnit] = useState(defaultTextEdit);
-        let [currEditName, setCurrentEditName] = useState({editingIndex:-1, selectedIndex: -1});
-
-
         if (isEditing) {
             return (
-                <Layout style={{padding: 10, paddingTop: 0}}>
-                    {/*Loops through all ingredients and renders them*/}
-                    {newIngredients.map((ingredient, index) => {
-                        return (
-                            <Layout key={index} style={{flexDirection: 'column', paddingBottom: 5}}>
-                                {/* Inputfield for ingredient name */}
-                                <Input style={{flex: 1}} label='Ingredient'
-                                    value={currEditName.editingIndex === index ? currEditName.text : ingredient.name}
-                                    onChangeText={text => setCurrentEditName({text,editingIndex:index})}
-                                    // The input field is in focus
-                                    onFocus={() => setCurrentEditName({editingIndex: index, text: ingredient.name})}
-                                    // The input lost focus
-                                    onBlur={() => {
-                                        // Update newIngredients with the new value
-                                        updateIngredientsArray(index, 'name', currEditName.text);
-                                        setCurrentEditName(defaultTextEdit)
-                                    }}
-                                ></Input>
-                                <Layout style={{flexDirection: 'row', justifyContent: 'center'}}>
-                                    {/* Inputfield for ingredient amount */}
-                                    <Input style={{flex: 1}} label='Mängd'
-                                        value={currEditAmount.editingIndex === index ? currEditAmount.text : (ingredient.amount ? String(ingredient.amount) : '')} 
-                                        onChangeText={text => setCurrentEditAmount({text, editingIndex:index})}
-                                        // The input field is in focus
-                                        onFocus={() => setCurrentEditAmount({editingIndex: index, text: (ingredient.amount ? String(ingredient.amount) : '')})}
-                                        // The input lost focus
-                                        onBlur={() => {
-                                            // Update newIngredients with the new value
-                                            updateIngredientsArray(index, 'amount', Number(currEditAmount.text));
-                                            setCurrentEditAmount(defaultTextEdit);
-                                        }}
-                                    ></Input>
-
-                                    {/* The unit selection is rendered as a dropdown */}
-                                    <Select style={{flex: 1, paddingLeft: 6, paddingRight: 6}} label='Enhet'
-                                        value={currEditUnit.editingIndex === index ? units[currEditUnit.selectedIndex] : ingredient.unit}
-                                        onSelect={(unit) => {
-                                            updateIngredientsArray(index, 'unit', units[unit.row]);
-                                            setCurrentEditUnit({editingIndex: -1, selectedIndex: -1});
-                                        }}>
-
-                                        {units.map((unit, index) => {
-                                            return (
-                                                <SelectItem key={index} title={unit} />
-                                            );
-                                        })}
-                                    </Select>
-                                    <Button 
-                                        style={{flex: 1, margin: 15}} 
-                                        status='danger' 
-                                        size='small' 
-                                        accessoryLeft={<Icon name='trash-2-outline' style={{width: 42, height: 42, marginLeft: 15}}/>}
-                                        >Ta bort</Button>
-                                </Layout>
-                                <Divider style={{backgroundColor: themeStyle['color-basic-600'], height: 3}}></Divider>
-                            </Layout>
-                        );
-                    })}
-                    <Layout style={{flex: 1, flexDirection: 'row'}}>
-                        <Button style={{margin: 5}}>Lägg till</Button>
-                    </Layout>
-                    
-                </Layout>
+                <EditIngredients newIngredientsState={newIngredientsState}></EditIngredients>
             );
         } else {
             return (
@@ -179,17 +104,6 @@ function FoodDetail({route, navigation}) {
                 </Layout>
             )
         }
-
-        // Updates the ingredient amount in newIngredients
-        function updateIngredientsArray(index, key, value) {
-            setNewIngredients(newIngredients.map((ingredient, i) => {
-                if (i === index) {
-                    ingredient[key] = value;
-                }
-                return ingredient;
-            }));
-        }
-        
     }
 
     return (
